requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
caesarcipher==1.0
chardet==3.0.4
discord==1.0.1
discord.py==1.3.2
dnspython==1.16.0
dnspython3==1.15.0
idna==2.9
idna-ssl==1.1.0
inflection==0.2.0
Mongothon3==0.7.20
multidict==4.7.5
pymongo==3.10.1
schedule==0.6.0
Schemer3==0.2.11.post1
typing-extensions==3.7.4.1
websockets==8.1
yarl==1.4.2
